requirements.txt
langchain
chainlit
langchain-community

python-dotenv
qdrant-client[fastembed]
pypdf